+++ SecureSoftware/Components/Pages/SecureNotePage.cs
@@ -1,5 +1,6 @@
 ﻿using SecureSoftware.Classes;
 using SecureSoftware.Forms;
+using SecureSoftware.Forms.Dialogs;
 
 namespace SecureSoftware.Components
 {
@@ -78,10 +79,20 @@ namespace SecureSoftware.Components
         async private void DeleteNotesButton_Click(object sender, EventArgs e)
         {
             SetActionRowEnabled(false);
-            bool isDeleted = await User.DeleteNotesAsync();
-            if (isDeleted)
+
+            PromptDeletion confirmationForm = new("Are you sure you want to delete all of your Secure Notes?", "Yes, delete all of them!", "No, keep them!");
+            DialogResult results = confirmationForm.ShowDialog();
+            if (results == DialogResult.OK)
             {
-                MainPanel.Controls.Clear();
+                bool isDeleted = await User.DeleteNotesAsync();
+                if (isDeleted)
+                {
+                    MainPanel.Controls.Clear();
+                }
+            }
+            else
+            {
+                confirmationForm.Dispose();
             }
             SetActionRowEnabled(true);
         }
